Refined the coding of initial transformation
Separated aesthetic curations out of init_special, for a better logistic.
Thus, the transformations for initial, medial, and final are all supposed to be bearing 3 parts in order: 1. the aesthetic curating phase, such as adjusting the position of dots of NE, NU. 2. the transformations for special combinations, such as BA, BO. 3. general transformations, such as C, J