Realize the umlaut vowels
Using a new strategy, that simply convert FVS2 to 'umlaut' which appears like a 0-width mark adjunct to the preceding vowel. The only additional work is then add 'umlaut' into the context of I.->aI.
Nevertheless, the position of 'umlaut' need to be adjusted for better looks in the final phase.