Separate the Urga style into a new project
Which was originally purposed for the italic style of Folon. But it turns out that they cannot share many OpenType Features in details, such as decofliction should be applied upon a larger scope since Lal is longer in Urga than in Folon, and etc. Besides, it's glories deserve an own font